+++ tests/client_tests/quote_post.py
@@ -9,13 +9,14 @@ with open(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data', 'ses
     session_info = json.load(f)
 
 response = request(
-    client_info["endpoint"], "timeline.get_post",
+    client_info["endpoint"], "timeline.quote_post",
      session_identification={
         "session_id": session_info["id"],
         "client_public_hash": client_info["public_hash"],
         "challenge_answer": answer_challenge(client_info["private_hash"], session_info["challenge"])
      },
-     post_id="e3c158bfc9495b94f7f768662b1cea58-9dd16ed2-1ef7-11ec-9a44-adbca69d60b4"
+     post_id="e3c158bfc9495b94f7f768662b1cea58-2079a68a-1ef6-11ec-83fe-731123d16b37",
+     text="This is a quoted post!"
 )
 
 print(json.dumps(response.data.result))